tidb: Move some test from tidb_test.go to session_test.go
tidb_test.go is too large now.
